backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework_simplejwt==5.3.1
django-cors-headers==3.5.0
pillow==10.3.0
django-filter==24.2